+++ src/RideWindow/ViewportArray.ts
@@ -1,0 +1,10 @@
+export class ViewportArray {
+     coordsXYZ: CoordsXYZ
+
+    stationName: string
+
+    constructor(coordsXYZ: CoordsXYZ, stationName: string) {
+        this.coordsXYZ = coordsXYZ
+        this.stationName = stationName
+    }
+}