+++ src/plugin.ts
@@ -4,7 +4,7 @@ import { startup } from "./startup";
 
 registerPlugin({
 	name: "EnhancedWindow",
-	version: "1.1.0",
+	version: "2.0.0",
 	authors: ["SkylineIndustries"],
 	type: "remote",
 	licence: "MIT",